+++ bluepysnap/circuit_ids.py
@@ -57,8 +57,11 @@ class CircuitIds(abc.ABC):
 
     @property
     def index_schema(self):
-        """Return an empty index with the same names of the wrapped index."""
-        return pd.MultiIndex.from_tuples([], names=self.index.names)
+        """Return an empty index with the same names and dtypes of the wrapped index."""
+        # NOTE: Since pandas 2.1.0, the index needs to contain the explicit dtypes. In pd.concat,
+        # the dtypes of multi-index are coerced to 'object' if any dataframe has indices with
+        # dtype='object'
+        return self.index[:0]
 
     @classmethod
     def _instance(cls, index, sort_index=True):